pyproject.toml
[tool.poetry]
name = "htcpcp-server"
version = "0.1.0"
description = ""
authors = ["impalah <impalah@gmail.com>"]
readme = "README.md"
packages = [{include = "htcpcp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.105.0"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
loguru = "^0.7.2"
mangum = "^0.19.0"
svix-ksuid = "^0.6.2"
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-instrumentation = "^0.48b0"
opentelemetry-exporter-otlp = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.23.3"
mock = "^5.1.0"
pytest-cov = "^4.1.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
files = ["src/"]  # Source code folder
ignore_missing_imports = false
strict = true
disallow_untyped_defs = false

